move metaclass into its own file, handle python 2.7 classmethod compatibility
