+++ dbt/adapters/base/__init__.py
@@ -1,3 +1,4 @@
+from dbt.adapters.base.meta import available
 from dbt.adapters.base.relation import BaseRelation
 from dbt.adapters.base.connections import BaseConnectionManager
-from dbt.adapters.base.impl import BaseAdapter, available
+from dbt.adapters.base.impl import BaseAdapter
