+++ dbt/compat.py
@@ -58,7 +58,16 @@ def write_file(path, s):
 
 
 if WHICH_PYTHON == 2:
-    # python 2.7 is missing this.
+    # In python 2, classmethod and staticmethod do not allow setters, so you
+    # can't treat classmethods as first-class objects like you can regular
+    # functions. This rarely matters, but for metaclass shenanigans on the
+    # adapter we do want to set attributes on classmethods.
+    class _classmethod(classmethod):
+        pass
+
+    classmethod = _classmethod
+
+    # python 2.7 is missing this
     class abstractclassmethod(classmethod):
         __isabstractmethod__ = True
 
@@ -72,9 +81,11 @@ if WHICH_PYTHON == 2:
         def __init__(self, func):
             func.__isabstractmethod__ = True
             super(abstractstaticmethod, self).__init__(func)
+
 else:
     abstractclassmethod = abc.abstractclassmethod
     abstractstaticmethod = abc.abstractstaticmethod
+    classmethod = classmethod
 
 
 def suppress_warnings():
